+++ src/test/e2e/commands/open-cache-file.test.ts
@@ -1,8 +1,9 @@
 import assert from "assert";
 import { LiquibaseGUITestUtils } from "../LiquibaseGUITestUtils";
-import { TextEditor } from "vscode-extension-tester";
+import { TextEditor, VSBrowser } from "vscode-extension-tester";
 import { DockerTestUtils } from "../../suite/DockerTestUtils";
 import { ContextOptions } from "../../../constants";
+import { Connection, ContextSelection } from "../../../cache";
 
 /**
  * Tests the opening of the cache.
@@ -51,8 +52,21 @@ suite("open cache", () => {
    * Tests that the cache files can be opened.
    */
   test("should open cache file", async () => {
+    const expectedConnectionFromCache: Connection = {
+      changelogs: [
+        {
+          path: LiquibaseGUITestUtils.CHANGELOG_FILE.toLowerCase(),
+          lastUsed: 1,
+          contexts: {
+            loadedContexts: ["bar", "baz", "foo"],
+            selectedContexts: ["foo"],
+          },
+        },
+      ],
+    };
+
     // create some data for the cache
-    await LiquibaseGUITestUtils.executeUpdate(configurationName, ContextOptions.LOAD_ALL_CONTEXT);
+    await LiquibaseGUITestUtils.executeUpdate(configurationName, ContextOptions.LOAD_ALL_CONTEXT, "foo");
 
     // open the cache file
     await LiquibaseGUITestUtils.startCommandExecution({
@@ -69,17 +83,52 @@ suite("open cache", () => {
     const key = keys[0];
 
     // get the cache for our key and check that the contexts are there
-    const cacheForKey = cache[key];
+    const cacheForKey: Connection = cache[key];
     // sanitize the result to remove the lastUsed from the changelogs
-    const sanitizedResult = {
-      contexts: cacheForKey.contexts,
-      changelogs: cacheForKey.changelogs.map((changelog: { path: string }) => ({
-        path: changelog.path.toLowerCase(),
-      })),
-    };
-    assert.deepStrictEqual(sanitizedResult, {
-      contexts: ["bar", "baz", "foo"],
-      changelogs: [{ path: LiquibaseGUITestUtils.CHANGELOG_FILE.toLowerCase() }],
+    cacheForKey.changelogs.forEach((pChangelog) => {
+      pChangelog.lastUsed = 1;
     });
+
+    assert.deepStrictEqual(cacheForKey, expectedConnectionFromCache);
+
+    //// now, check that the selection is there in the context input
+    const input = await LiquibaseGUITestUtils.startCommandExecution({
+      command: "update...",
+      configurationName,
+      changelogFile: true,
+    });
+
+    await input.setText(ContextOptions.USE_RECENTLY_LOADED);
+    await input.confirm();
+    await VSBrowser.instance.takeScreenshot("00");
+
+    await LiquibaseGUITestUtils.waitForCheckboxesToBeThere(input);
+    const quickPicks = await input.getCheckboxes();
+
+    const actualContexts: ContextSelection = {
+      loadedContexts: [],
+      selectedContexts: [],
+    };
+
+    // check every quick pick for the cached selection
+    for (const quickPick of quickPicks) {
+      const label = await quickPick.getLabel();
+      const selected = await quickPick.getAttribute("aria-checked");
+
+      actualContexts.loadedContexts?.push(label);
+
+      if (selected === "true") {
+        actualContexts.selectedContexts?.push(label);
+      }
+    }
+
+    assert.deepStrictEqual(
+      actualContexts,
+      expectedConnectionFromCache.changelogs[0].contexts,
+      "contexts with selections"
+    );
+
+    // cancel any input after the test
+    await input.cancel();
   });
 });